rearranging code structure to adhere to docker-compose in project dir and dockerfile in subdir company wide standard; also adding ray+fastapi deployment; removing extraneous files copied into container workdir
